web: combine score_details blueprint into views
