--- depobs/website/score_details/blueprint.py
@@ -1,128 +1,0 @@
-import logging
-from typing import Dict, Optional, Type
-
-from flask import Blueprint
-import graphviz
-import networkx as nx
-from werkzeug.exceptions import NotFound
-
-from depobs.database import models
-from depobs.scanner import graph_traversal
-from depobs.scanner import graph_util
-from depobs.worker import scoring
-
-
-log = logging.getLogger(__name__)
-
-score_details_blueprint = api = Blueprint(
-    "score_details_blueprint", __name__, url_prefix="/score_details",
-)
-
-
-@api.route("/graphs/<int:graph_id>", methods=["GET"])
-def get_graph(graph_id):
-    """
-    Returns an svg rendered graphviz dot graph of the given scan
-
-    graph result for the given graph_id
-    """
-    db_graph: models.PackageGraph = models.get_graph_by_id(graph_id)
-    g: nx.DiGraph = graph_util.package_graph_to_networkx_graph(db_graph)
-    graph_util.update_node_attrs(
-        g,
-        label={
-            pv.id: f"{pv.name}@{pv.version}"
-            for pv in db_graph.distinct_package_versions_by_id.values()
-        },
-    )
-    dot_graph: graphviz.Digraph = graph_util.nx_digraph_to_graphviz_digraph(g)
-    dot_graph.attr(rankdir="LR")
-
-    log.debug(f"dot for graph {graph_id}: {dot_graph!r}")
-    return dot_graph.pipe(format="svg").decode("utf-8")
-
-
-@api.route("/condensate_graphs/<int:graph_id>", methods=["GET"])
-def get_condensate_graph(graph_id):
-    """
-    Returns an svg rendered graphviz dot graph of the given scan
-
-    graph result for the given graph_id
-    """
-    db_graph: models.PackageGraph = models.get_graph_by_id(graph_id)
-    g: nx.DiGraph = graph_util.package_graph_to_networkx_graph(db_graph)
-    c: nx.DiGraph = graph_traversal.condensation(g)
-
-    graph_util.update_node_attrs(
-        g,
-        label={
-            pv.id: f"{pv.name}@{pv.version}"
-            for pv in db_graph.distinct_package_versions_by_id.values()
-        },
-    )
-    dot_graph: graphviz.Digraph = graph_util.nx_digraph_to_graphviz_digraph(c)
-    dot_graph.attr(rankdir="LR")
-
-    for scc_node_ids in graph_traversal.outer_in_dag_iter(c):
-        with dot_graph.subgraph() as s:
-            s.attr(rank="same")
-            for scc_node_id in scc_node_ids:
-                log.debug(
-                    f"scc_node_id {scc_node_id} members {c.nodes[scc_node_id]['members']}"
-                )
-                component_pkgs = "\n".join(
-                    g.nodes[m_id]["label"] for m_id in c.nodes[scc_node_id]["members"]
-                )
-                s.node(
-                    str(scc_node_id),
-                    label=f"scc_id: {scc_node_id}\nmembers: {component_pkgs}",
-                )
-
-    log.debug(f"dot for condensate graph {graph_id}: {dot_graph!r}")
-    return dot_graph.pipe(format="svg").decode("utf-8")
-
-
-@api.route(
-    "/score_component_graph/<int:graph_id>/<string:package_report_field>",
-    methods=["GET"],
-)
-def get_scoring_graph(graph_id: int, package_report_field: str):
-    log.info(
-        f"getting scoring graph {graph_id} for package report field {package_report_field}"
-    )
-    component: Optional[
-        Type[scoring.ScoreComponent]
-    ] = scoring.find_component_with_package_report_field(package_report_field)
-    if component is None:
-        return NotFound(
-            f"Unable to find scoring component to provide field {package_report_field}"
-        )
-    log.info(f"rendering score component graph with component: {component}")
-
-    # find graph and score it with that component
-    db_graph: models.PackageGraph = models.get_graph_by_id(graph_id)
-    g: nx.DiGraph = graph_util.package_graph_to_networkx_graph(db_graph)
-
-    reports_by_package_version_id: Dict[
-        models.PackageVersionID, models.PackageReport
-    ] = scoring.score_package_graph(db_graph, [component], g)
-
-    graph_util.update_node_attrs(
-        g,
-        report=reports_by_package_version_id,
-        label={
-            pv.id: f"{pv.name}@{pv.version}"
-            for pv in db_graph.distinct_package_versions_by_id.values()
-        },
-    )
-
-    dot_graph: graphviz.Digraph = graph_util.nx_digraph_to_graphviz_digraph(g)
-    dot_graph.attr(rankdir="LR")
-
-    for node_id, node_data in g.nodes(data=True):
-        # add package report field name and value to the label
-        label = f"{node_data['label']}\n\n{package_report_field}:\n{getattr(node_data['report'], package_report_field, None)}"
-        dot_graph.node(str(node_id), label=label)
-
-    log.debug(f"rendering score component graph: {dot_graph!r}")
-    return dot_graph.pipe(format="svg").decode("utf-8")
